calculated statuses, removed calculated scores
Big Update!

Previously, employees and reviewers would reach out to HR dozens of
times during the performance review period, confused about how to
proceed to the next step.

This update introduces a few major usabiliy improvements that should
reduce support requests dramatically, saving HR dozens of hours of work
during the review period, and helping employees and reviewers to
complete their reviews more quickly.

We now show status messages that tell employees and reviewers how many
more updates they need to make before they can move to the next step,
and what sort of updates are needed. For example, maybe the employee
didn't add all six new goals, or the reviewer forgot to review a past
goal. Now, there's a clear status message in each major tab, and it
explains what's missing.

We also highlighted the "Next" and "Done" buttons, to make it very
clearly what people need to click when they've completed enough of the
review to move to the next step.

Crazy technical error messages no longer appear if information is
missing. All of the status messages are friendly now, and they explain
what needs to be done next.

We now prevent people from accidentally (or intentionally) deleting
important information from the review if it was reuired in a previous
step.

Lastly, per HR request, we removed the calculate score code due to a
change to in the way the reviewer scores employees. (Scoring is *much*
simpler now.) The old Java file that handled score calculation is still
in the repository for reference, renamed to
"inactive-calculate-scores.java". It's still in the repository in case
there's a future change where a calculated score would be useful again.

Hope that helps!